requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
streamlit>=1.28.0
streamlit-chat>=0.1.1
gradio>=3.50.0
pandas>=1.5.0
openpyxl>=3.0.0